+++ src/yafowil/widget/dict/resources/widget.js
@@ -12,14 +12,14 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
     $(document).ready(function() {
         // initial binding
         yafowil.dict.binder();
-        
+
         // add after ajax binding if bdajax present
         if (typeof(window['bdajax']) != "undefined") {
             $.extend(bdajax.binders, {
                 dictwidget_binder: yafowil.dict.binder
             });
         }
-        
+
         // add binder to yafowil.widget.array hooks
         if (typeof(window.yafowil['array']) != "undefined") {
             $.extend(yafowil.array.hooks.add, {
@@ -27,48 +27,48 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
             });
         }
     });
-    
+
     $.extend(yafowil, {
-        
+
         dict: {
-            
-            create_row: function() {
+
+            create_row: function(key_css, val_css) {
                 var row = '';
                 row += '<tr>';
                 row +=   '<td class="key">';
-                row +=     '<input type="text" value="" />';
+                row +=     '<input type="text" class="' + key_css + '" value="" />';
                 row +=   '</td>';
                 row +=   '<td class="value">';
-                row +=     '<input type="text" value="" />';
+                row +=     '<input type="text" class="' + val_css + '" value="" />';
                 row +=   '</td>';
-                row +=   '<td>';
+                row +=   '<td class="actions">';
                 row +=     '<div class="dict_actions">';
                 row +=       '<a class="dict_row_add" href="#">';
-                row +=         '<i class="icon-plus-sign">&nbsp;</i>';
+                row +=         '<span class="icon-plus-sign"> </span>';
                 row +=       '</a>';
                 row +=       '<a class="dict_row_remove" href="#">';
-                row +=         '<i class="icon-minus-sign">&nbsp;</i>';
+                row +=         '<span class="icon-minus-sign"> </span>';
                 row +=       '</a>';
                 row +=       '<a class="dict_row_up" href="#">';
-                row +=         '<i class="icon-circle-arrow-up">&nbsp;</i>';
+                row +=         '<span class="icon-circle-arrow-up"> </span>';
                 row +=       '</a>';
                 row +=       '<a class="dict_row_down" href="#">';
-                row +=         '<i class="icon-circle-arrow-down">&nbsp;</i>';
+                row +=         '<span class="icon-circle-arrow-down"> </span>';
                 row +=       '</a>';
                 row +=     '</div>';
                 row +=   '</td>';
                 row += '</tr>';
                 return row;
             },
-            
+
             get_row: function(action) {
                 return $(action).parent().parent().parent();
             },
-            
+
             base_name: function(context) {
                 return context.parents('.dictwidget').attr('id');
             },
-            
+
             reset_indices: function(context) {
                 var index = 0;
                 var base_name = yafowil.dict.base_name(context);
@@ -88,7 +88,7 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
                 });
                 yafowil.dict.binder(context);
             },
-            
+
             mark_disabled: function(context) {
                 context = $(context);
                 $('a.dict_row_up', context)
@@ -100,32 +100,42 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
                     .last()
                     .addClass('dict_row_down_disabled');
             },
-            
+
+            add_class: function(trigger, name) {
+                var table = $('table', $(trigger).parents());
+                classes = table.attr('class').split(' ');
+                var idx, css;
+                for (idx in classes) {
+                    css = classes[idx];
+                    if (css.substring(0, name.length) == name) {
+                        return css.substring(name.length + 1, css.length);
+                    }
+                }
+                return '';
+            },
+
             binder: function(context) {
                 yafowil.dict.mark_disabled(context);
                 $('a.dict_row_add', context)
                     .unbind()
                     .bind('click', function(event) {
                         event.preventDefault();
+                        var key_css = yafowil.dict.add_class(this, 'key');
+                        var val_css = yafowil.dict.add_class(this, 'value');
                         var row = yafowil.dict.get_row(this);
-                        var new_row = yafowil.dict.create_row();
+                        var new_row = yafowil.dict.create_row(key_css, val_css);
                         var container = row.parent();
                         if (container.get(0).tagName.toLowerCase() == 'tbody') {
                             row.after(new_row);
                         } else {
                             var table = container.parent();
                             var body = $('tbody', table);
-                            // jq 1.4.4 fails atm
-                            //if (!body.length) {
-                            //    body = $('<tbody />');
-                            //    table.append(body);
-                            //}
                             container = body;
                             container.prepend(new_row);
                         }
                         yafowil.dict.reset_indices(container);
                     });
-                
+
                 $('a.dict_row_remove', context)
                     .unbind()
                     .bind('click', function(event) {
@@ -135,7 +145,7 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
                         row.remove();
                         yafowil.dict.reset_indices(container);
                     });
-                
+
                 $('a.dict_row_up', context)
                     .unbind()
                     .bind('click', function(event) {
@@ -144,7 +154,7 @@ if (typeof(window['yafowil']) == "undefined") yafowil = {};
                         row.insertBefore(row.prev());
                         yafowil.dict.reset_indices(row.parent());
                     });
-                
+
                 $('a.dict_row_down', context)
                     .unbind()
                     .bind('click', function(event) {